Refactor context providers for lint compatibility
