+++ src/hooks/auth-context.ts
@@ -1,0 +1,11 @@
+import { createContext } from 'react';
+import type { Session, User } from '@supabase/supabase-js';
+
+export interface AuthContextType {
+  user: User | null;
+  session: Session | null;
+  isLoading: boolean;
+  signOut: () => Promise<void>;
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
