+++ src/hooks/useAuth.ts
@@ -1,0 +1,13 @@
+import { useContext } from 'react';
+import { AuthContext } from './auth-context';
+
+/**
+ * Access the authentication context.
+ */
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error('useAuth must be used within an AuthProvider');
+  }
+  return context;
+};
